Refactor files, move executable to bin
