+++ src/get-files-by-pattern.ts
